chore(tanstack-query-react-plugin): move files and update snapshots
